// src/types/printFormTypes.ts
export type PrintPageHeaderConfig = {
  revision?: string | number;
  number?: string | number;
  prefix?: string | number;
  title?: string;
  documentType?: string;
  date?: string | number;
};